requirements.txt
# streamlit==1.32.0
# python-dotenv==1.0.0
# sentence-transformers==2.3.1
# PyPDF2==3.0.1
# python-docx==1.0.1
# requests==2.31.0
# beautifulsoup4==4.12.3
# faiss-cpu==1.7.4
# groq==0.4.1
# numpy==1.24.4
# supabase==1.0.3
# werkzeug==2.3.6
# postgrest==0.10.6

streamlit==1.32.0
python-dotenv==1.0.0
sentence-transformers==2.3.1
PyPDF2==3.0.1
python-docx==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
faiss-cpu==1.7.4
groq==0.4.1
numpy==1.24.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pyjwt == 2.9.0
